formatting, refactoring out drp status endpoint into drp-specific endpoint
